convert tslCertificate component to rfc
